// src/app/core/models/email.ts
import { SelectOption } from "./select-option";

export class Email {

    subject: string;
    from: string;
    to: string;
    date: Date;
    size: number;
    uid: number;
    msgno: number;
    recent: number;
    flagged: number;
    answered: number;
    deleted: number;
    seen: number;
    draft: number;
    udate:number;

}